Removed needless model reuse, reorganised
